feat: add max_steps parameter

+++ model_training_utils.py
@@ -172,7 +172,7 @@ def run_customized_training_loop(
     raise ValueError(
         'if `metric_fn` is specified, metric_fn must be a callable.')
 
-  total_training_steps = steps_per_epoch * epochs
+  total_training_steps = int(steps_per_epoch * epochs)
 
   # To reduce unnecessary send/receive input pipeline operation, we place input
   # pipeline ops in worker task.
